Lab2_Binary_Semantic_Segmentation_2025/requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
filelock==3.18.0
fsspec==2025.3.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pytorch-ranger==0.1.1
PyYAML==6.0.2
scipy==1.15.2
simsimd==6.2.1
stringzilla==3.12.3
sympy==1.13.1
torch==2.6.0
torch-optimizer==0.3.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
